Move EntityEvolver components to modelling
Move EntityEvolver components to modelling. The EntityEvolver and
PayloadBasedEntityEvolver are not event sourcing specifics after all.
We'd ideally also move the AnnotationBasedEntityEvolver, but this
current only depends on the EventSourcingHandler annotation. As such,
it's still in the event sourcing package
