+++ eventsourcing/src/main/java/org/axonframework/eventsourcing/SimpleEventSourcedComponent.java
@@ -22,6 +22,7 @@ import org.axonframework.common.infra.DescribableComponent;
 import org.axonframework.eventhandling.EventMessage;
 import org.axonframework.messaging.QualifiedName;
 import org.axonframework.messaging.unitofwork.ProcessingContext;
+import org.axonframework.modelling.EntityEvolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
