+++ eventsourcing/src/main/java/org/axonframework/eventsourcing/configuration/DefaultEventSourcedEntityBuilder.java
@@ -21,8 +21,8 @@ import org.axonframework.configuration.ComponentFactory;
 import org.axonframework.configuration.Configuration;
 import org.axonframework.eventsourcing.EventSourcingRepository;
 import org.axonframework.eventsourcing.CriteriaResolver;
-import org.axonframework.eventsourcing.EntityEvolver;
-import org.axonframework.eventsourcing.PayloadBasedEntityEvolver;
+import org.axonframework.modelling.EntityEvolver;
+import org.axonframework.modelling.PayloadBasedEntityEvolver;
 import org.axonframework.eventsourcing.SimpleEventSourcedComponent;
 import org.axonframework.eventsourcing.annotation.EventSourcedEntityFactory;
 import org.axonframework.eventsourcing.eventstore.EventStore;
