+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/configuration/DefaultEventSourcedEntityBuilderTest.java
@@ -20,7 +20,7 @@ import org.axonframework.common.infra.MockComponentDescriptor;
 import org.axonframework.eventhandling.GenericEventMessage;
 import org.axonframework.eventsourcing.EventSourcingRepository;
 import org.axonframework.eventsourcing.CriteriaResolver;
-import org.axonframework.eventsourcing.EntityEvolver;
+import org.axonframework.modelling.EntityEvolver;
 import org.axonframework.eventsourcing.SimpleEventSourcedComponent;
 import org.axonframework.eventsourcing.annotation.EventSourcedEntityFactory;
 import org.axonframework.eventsourcing.eventstore.EventCriteria;
